feat(hiring): agrrgar formulario de nueva convocatoria

+++ frontend/src/components/QuickSearchInput.tsx/QuickSearchInput.tsx
@@ -4,7 +4,8 @@ import debounce from 'debounce'
 import { Loader2, X } from 'lucide-react'
 import { Card } from '../ui/card'
 import { Button } from '../ui/button'
-import { useState } from 'react'
+import { useState, useRef, useEffect } from 'react'
+import { ScrollArea } from '../ui/scroll-area'
 
 interface Props<T> {
   placeholder?: string
@@ -14,6 +15,7 @@ interface Props<T> {
   renderSelected: (item: T) => React.ReactNode
   className?: string
   defaultValue?: keyof T
+  notKeepSelected?: boolean
 }
 
 export default function QuickSearchInput<T>({
@@ -23,26 +25,64 @@ export default function QuickSearchInput<T>({
   renderOption,
   renderSelected,
   className,
+  notKeepSelected = false,
 }: Props<T>) {
   const [item, setItem] = useState<T | null>(null)
+  const [inputValue, setInputValue] = useState<string>('') // Estado para el valor del input
+  const [isOpen, setIsOpen] = useState(false)
+  const containerRef = useRef<HTMLDivElement>(null)
+
   const { mutate, isPending, data, isSuccess } = useMutation({
     mutationFn: searchFn,
   })
+
   const onSelect = (item: T) => {
-    setItem(item)
+    if (!notKeepSelected) {
+      setItem(item)
+    } else {
+      setInputValue('') // Limpia el input si no se mantiene la selección
+    }
+    setIsOpen(false)
     handleSelect(item)
   }
+
   const onRemove = () => {
     setItem(null)
+    setInputValue('')
+    setIsOpen(false)
     handleSelect(null)
   }
-  const requested = isPending || isSuccess
-  const onChangeInput = debounce((e: React.ChangeEvent<HTMLInputElement>) => {
+
+  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     const query = e.target.value || ''
+    setInputValue(query) // Actualiza el estado del input inmediatamente
+    debouncedSearch(query) // Ejecuta la búsqueda con retraso
+  }
+
+  // Función de búsqueda con debounce aplicada
+  const debouncedSearch = debounce((query: string) => {
     mutate(query)
+    setIsOpen(true) // Abre la lista al empezar a escribir
   }, 300)
+
+  useEffect(() => {
+    const handleClickOutside = (event: MouseEvent) => {
+      if (
+        containerRef.current &&
+        !containerRef.current.contains(event.target as Node)
+      ) {
+        setIsOpen(false) // Cierra la lista si haces clic fuera del contenedor
+      }
+    }
+
+    document.addEventListener('mousedown', handleClickOutside)
+    return () => {
+      document.removeEventListener('mousedown', handleClickOutside)
+    }
+  }, [])
+
   return (
-    <div className={`relative ${className}`}>
+    <div ref={containerRef} className={`relative ${className}`}>
       {item ? (
         <Card className="p-2 flex items-center justify-between">
           {renderSelected(item)}
@@ -58,30 +98,35 @@ export default function QuickSearchInput<T>({
         </Card>
       ) : (
         <Input
-          onChange={onChangeInput}
+          onChange={handleInputChange} // Usa la función sin debounce para actualizar el input
+          value={inputValue} // Asigna el valor del input desde el estado
           placeholder={placeholder}
           type="search"
         />
       )}
-      {requested && !item && (
-        <ul className="absolute z-10 w-full mt-1 border rounded-md max-h-60 overflow-auto bg-primary-foreground">
-          {isPending && (
-            <li className="w-full grid place-content-center">
-              <Loader2 className="animate-spin" />
-            </li>
-          )}
-          {isSuccess &&
-            data?.map((item: T, idx) => (
-              <li
-                key={idx}
-                className="w-full px-1"
-                onClick={() => onSelect(item)}
-              >
-                {renderOption(item)}
-              </li>
-            ))}
-        </ul>
-      )}
+      <div className="absolute">
+        {isOpen && !item && (
+          <ScrollArea className="max-h-52 z-10 w-full mt-1 border rounded-md bg-background flex-col">
+            <ul>
+              {isPending && (
+                <li className="max-w-full py-1 pl-4 text-sm hover:bg-muted rounded-sm px-1">
+                  <Loader2 className="animate-spin" />
+                </li>
+              )}
+              {isSuccess &&
+                data?.map((item: T, idx) => (
+                  <li
+                    key={idx}
+                    className="w-full py-1 pl-4 text-sm hover:bg-muted rounded-sm mx-1"
+                    onClick={() => onSelect(item)}
+                  >
+                    {renderOption(item)}
+                  </li>
+                ))}
+            </ul>
+          </ScrollArea>
+        )}
+      </div>
     </div>
   )
 }
